+++ tree/include/Tree.h
@@ -9,7 +9,7 @@ public:
     Tree(std::shared_ptr<Board> currentPosition);
     ~Tree();
     void generateTree(int maxDepth);
-    float getPositionValue();
+    int getPositionValue();
     void optimalSequence();
     int totalNodesGenerated();
