+++ tree/src/Evaluator.cpp
@@ -6,8 +6,8 @@
 #include "../include/Evaluator.h"
 #include <iostream>
 
-float Evaluator::evaluatePosition(Board &position) {
-    float positionValue = 0; 
+int Evaluator::evaluatePosition(Board &position) {
+    int positionValue = 0;
     if (position.checkmate()) {
         return position.whiteOrder() ? -MAX_POS_VAL : MAX_POS_VAL; // float isn't precise enough, so FLT_MAX - 1 == FLT_MAX 
     }
